Be more careful in intersection solver at start/end of segment 1
At ordinary intersections we use the intersection angle to decide on the
next step forward. This does not work for intersections at the start and
end points and we are falling back to finding inside/outside
numerically. However, this was only done for one of the segments,
leading to rendering bugs if it happened on the other segment. This
patch also applies the numerical method to the second segment.

Pick-to: 6.7
Fixes: QTBUG-125864
Change-Id: I490acb74ea6215efa1986c7c6bfbcdd8c3172120
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>