Ensure that "styleName" is set by Qt.font
Qt.font is a method on the global Qt object that allows creating a `font`
value type from a JS object.

The `QtObject::font` implementation for the method delegates to
`QQmlValueTypeProvider::createValueType(const QJSValue&, ...)`, which
takes care of setting up the relevant structure from the `QJSValue`
wrapping the JS object.

Currently, this will call the `create` function on
`QQuickFontValueType`, the value type wrapper for the `QFont` type,
which will extract all relevant properties and set them accordingly.

The `create` function went out of sync with the actual structure of the
type, so that it will not extract and set a `styleName` property from
the JS object even if it is exposed by the value type.

Thus, for example, trying to set `styleName` with call such as:

```
Qt.font({ styleName: "some style" })
```

Will produce a `font` type that has a default `styleName` instead of the
provided one.

Thus, the `create` method for `QQuickFontValueType`, was amended to take
into consideration the `styleName` property.

A new test-case was added to ensure that usages of `Qt.font` will set a
`styleName` when it is provided.

Fixes: QTBUG-125495
Change-Id: Ic048d18012370e59947cbd19d116039799446458
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>