dom: add dom representation for lambdas
Populate a FunctionExpression script element via GenericScriptElement
in qqmldomastcreator by reusing some code used by the
non-lambda methods by extracting it in to a "prepareBodyForFunction"
helper method.

The introduction of lambdas broke some assumptions in the
qqmldomastcreator, namely that the order in which some lists were
visited did not matter because they did not contain any qqmljsscopes.
Functionexpressions do contains a qqmljsccope for the scope they contain
and can appear everywhere expressions can appear, so fix the visitation
order for those lists.

Introduce a helper method endVisitForLists() that collects all the list
elements of a list and use it for the trivial lists. Also get rid of the
empty visit() methods for those lists.

Make qqmldomastcreator visit the FormalParameterLists so they can be
modelled in FunctionExpressions, and adapt the non-lambda method
construction to use the FormalParameterlists.

Set the function's qqmljsscope in FunctionExpression in
QQmlDomAstCreatorWithQQmlJSScope::setScopeInDomAfterEndvisit instead of
setting the function's parent scope.

Finally add some tests for the dom structure of the lambdas.

Also add some tests for generators (they are handled like lambdas for
now), which requires implementing the dom representation for yield
expressions.

Task-number: QTBUG-92876
Change-Id: I2fc37edfb7aadb169da7c50e3d0fc47182090569
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>