rhi: Skip using lockfile for cache when using QSaveFile
The RHI pipeline cache is saved to disk using either QSaveFile or QFile
depending on build flags. When using a regular file we need to have a
lock file so it is not used whilst we are writing.

With QSaveFile this is not necessary, it writes to a temporary file and
moves over atomically. The lock file is problematic as it causes a
fdatasync blocking until file changes are flushed to disk.

Change-Id: Ic333f05efc3c893750fcda746bf886b36a9bba18
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>