Don't update transform if it hasn't changed
Updating the transform marks it as dirty and causes rebatching.
Since we were doing this even when nothing had changed, we were
essentially rebatching every time there was any update to the
shape. We should leave it to the node to decide if it has
been updated.

This adds two tests that verify that rebatching still happens
when changing the color/gradient of the shape.

Change-Id: Iefb7a21a89ac2c31053e473cf689fa06e34a40d5
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit e3d0233bbe3e130c72e3957c45007097c019530d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>