Mark a possibly unusual pattern with a comment
Generally, we should avoid having unscoped values lying around as they
could be collected by the GC.

A pattern in `qqmlglobal.cpp` makes a direct cast to `QV4::value` that
is safe due to the value being tied to a `QJSValue`.

Add a comment to the two usages of the pattern to identify that using a
naked `Value` might not always be safe.

The added comment is intended to make the general issue with `Value`s
more apparent to people with less experience working with the code-base,
make the reason the usage is safe in this case more explicit for people
who have cursory knowledge of the possible issue and, possibly, help
thinking in more correct terms when similar patterns are written in
other parts of the code-base.

Change-Id: Ie2614079aaad896eb86d9491cdf0f2d61c321e75
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>