Tie the libraries for QtQuick, QtQml, Models, and WorkerScript together
When QtQuick is loaded, we know that we also need QtQml, models and
workerscript. Therefore, hold on to some symbols in those so as to
prevent the linker from optimizing the depedencies out and forcing the
engine to use plugins.

This doesn't work on QNX, for unknown reasons. Therefore, on QNX we keep
loading the plugin for WorkerScript.

Pick-to: 6.8
Change-Id: I3653aac46028521160e4f96274c90c586bde1512
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>