Fix table view to provide the right index to the section size providers
The reordering of sections updates the logical and visual index mapping
table internally, which is further used to provide the correct index to
the public APIs. However, this creates an issue when the index provided
to the columnWidthProvider() or rowHeightProvider() was already a
logical index and this once again converted in explicitColumnWidth() or
explicitRowHeight().

This patch fixes it by providing the column or row view index to the
section size providers.

Change-Id: If0be55b1039e5fcea7b366570da020ae5e32b517
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 558337a0a1b0ca20f248762165fc254b6cb8fa3a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>