StyleGenerator: Account for strokes that are drawn outside the atom
In Figma, the stroke can be drawn:

INSIDE: stroke drawn inside the shape boundary
OUTSIDE: stroke drawn outside the shape boundary
CENTER: stroke drawn centered along the shape boundary

So far we were only accounting for the case where the stroke was
drawn inside, and if the stroke would be outside, we wouldn't include
it in the geometry of the atom, but consider it part of the "shadow".
This is not correct, because in Quick Controls, the stroke, or border,
is part of the background of the control and drawn inside the render
bounds. Only shadows should be excluded from the geometry of the
control.

That is why we need to account for these in the stylegenerator when
exporting the geometry, calculating offsets and paddings.

Change-Id: Ia9184c5fa97d2002d2dafa7c33b8091836650889
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>