QQuickMenu: clear currentIndex upon a call to popup()
If you popup() a Menu a second time, without providing
a MenuItem as argument, the previously activated MenuItem
will still be current, and rendered as highlighted.
This is not the expected behavior for Context Menus
or MenuBar menus.

This patch will therefore make sure that we reset
currentIndex when popping up a Menu. It's still possible
to open() it instead, if currentIndex should be kept.

Change-Id: I09fdc2e5429ba107b236d0c761ca8d1de654710e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit e3bab6a76e751153eee7b661a910a0beb73c97e3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>