QQuickPopupPositioner: implement setEffectivePosFromWindowPos()
Factor out setting the effective position to a separate
function, and call it from all places where we change it.
This ensures that we always emit the xChanged/yChanged
signals, and that we always include any drop-shadow offset
in the calculation.
Also change the effectPos variable to be a QPointF.

Change-Id: I00ad7e25397887719811738343448f177261cc54
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 520cbb216b4adb56cdc701562a08ae5156f9101c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>