Flickable: Don't accept ScrollEnd wheel event
...to fix snapping behavior of
ListView { snapMode: SnapOneItem; delegate: ListView { ...} }

It doesn't accept the event with ScrollBegin phase, so shouldn't do it
for the ScrollEnd phase either.

Accepting ScrollEnd unconditionally broke nested ListView, for example,
where the higher-level one wouldn't snap back to the delgate and would
get stuck in a moving state -- that is, until you click it so that the
mouseReleaseEvent cleans it up.

Fixes: QTBUG-126042
Pick-to: 6.5 6.7 6.8
Change-Id: I94b0a3570e4750d68501095c56f06f5cca31923f
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>