qmltyperegistrar: Consider all relations for processed foreign types
If a type was added as an extension, attachment, or sequence value type,
the dependency's qmltypes file will already have it. There's no need to
duplicate it. This is most prominent in the case of the "Qt" namespace
which is an extension to the "Qt" object and shouldn't be repeated in
every single module that uses its enums.

Pick-to: 6.8
Change-Id: I9a309d8a6aae4299d1b537e2b12796f0839cd192
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>