qmlls: don't crash on missing logger
The logger pointer in qqmljstypresolver used by qmlls happens to point
to a stack variable, which leads to a use-after-free when the
typeresolver wants to complain about a missing module.
This leads to crashes when qtdeclarative is not completely built and
some modules are missing for example.

Replace the stack variable with an std::shared_ptr and make it part of
the typeresolver dependencies so it does not get destroyed while the
typeresolver has a reference to it.

Change-Id: I579c528943e011b2bc4c2b14235650dbe9ffeb74
Reviewed-by: Dmitrii Akshintsev <dmitrii.akshintsev@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 3b63aa3e57a4f5f7bac8d077d0a5b3dda2e26661)