gc: fix marking for (Weak)(Map|Set)
Map and Set must hold on to their keys (and values in case of Map),
which did not happen if we inserted values into an already marked Map or
Set.
A similar issue holds for weak maps and sets, except that for those we
only want to mark them once we are past the FreeWeak(Maps|Sets) phase.

Pick-to: 6.8
Change-Id: Ibc05be4a85172a9b726beebf0d008acde4479edf
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>