Disambiguate constant types and CONSTANT properties in qmltypes file
When writing a `qmltypes` file, `qmltyperegistrar` could write to the
same binding twice representing two different pieces of information.

A `Property` element in a `qmltypes` file, contains information about
the property itself and the type of the property, in the same scope.

When the type of a property is `const`, an `isConstant: true` binding is
written for the `isConstant` property of the `Property` element.

When the property is `CONSTANT`, then, similarly, and `isConstant: true`
binding is written for the `isConstant` property of the `Property`
element.

The usage of the same property makes it impossible to distinguish
between the two cases.

To avoid the issue, the code was modified to write a `isTypeConstant`
binding for the first case and a `isPropertyConstant` binding for the
second case.

The definition for the `Property` element was modified to represent the
new properties.

The definition for `Method` and `Parameter` were modified to so that
their `isConstant` property was renamed to `isTypeConstant`.
While they didn't present the same ambiguity, as they only contain type
information, the code that writes types information in the "qmltypes"
file is shared between them so that they required to be adapted to the
new naming.

`QQmlJSTypeDescriptionReader`, which can interpret the information
obtained from a `qmltypes` file, was generally modified to respect the
new naming, while keeping the behavior mostly intact.

Previously, the `isConstant` property in a `Property` element was
interpreted as referring to the property itself so that the type
constant information was lost or misinterpreted.

To make space for the previously lost information, `QQmlJSMetaProperty`
was modified to have an additional `isTypeConstant`.
For consistency, the pre-existing `m_isConstant` member was renamed to
`m_isPropertyConstant`.

`QQmlJSTypeDescriptionReader` will interpret both the `isTypeConstant`
and `isPropertyConstant` bindings, when reading a `Property` element,
and will reflect the information on the resulting `QQmlJSMetaProperty`.

Pre-existing test-cases in `tst_qmltyperegistrar` were adapted to the
changes.

One `qmltypes` file in the data corpus for `tst_qmllint` was modified to
respect the new naming.

Change-Id: I10897bb421b83b92f3a19082273b64a5c66adc07
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>