PopupPositioner: reset popupItem position when using popup window
If you switch from using Popup.Item to Popup.Window at run-time,
it might happen that the popupItem has an x and y position from
a previous time it was shown as Popup.Item. This patch will
therefore make sure that we move it back to 0, 0 when placed
inside a window.

Pick-to: 6.8
Change-Id: Ib11e209f3a14a93bb5f5f576005a28c3233f8342
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>