tst_combobox: Fix some failing tests for FluentWinUI3 style
- When checking scale animation transition, first make sure that the
transition includes an animation with the scale property.
The FluentWinUI3 style animates on a different property and the test
fails.
- Skip the test for the FluentWinUI3 style when checking that the y
coordinate is never changed between opening and reopening the popup.
This depends on the implicit height of the popup for the FluentWinUI3
style, and that changes between opening/reopening.
- Make sure that popup enter transitions are finished before testing
events

Pick-to: 6.8
Change-Id: Ic3ca570db5e4ce399f52d7d2f266a4ba48a255a6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>