QtQml: Fix misuse of logging categories
If we manually define the logging category function, we should really
avoid the macro for declaring it. When exporting a logging category, we
need to use the special export macro since we shouldn't rely on the
internals of the declaration macro for placing the export. The same
holds for static logging categories.

Furthermore, consistently call the functions for lcQml and lcJs.

We also need to fix the namespaces in qmlls now since the logging
categories are decidedly not in either of the used namespaces, and the
namespaces are all inside any Qt namespace.

Task-number: QTBUG-67692
Change-Id: I00fdcb80379e47dd7e928260e9c1dec91f22ad4b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>