Rename *isConstant for QQmlJSMetaProperty
`QQmlJSMetaProperty` exposes an set/`isConstant` interface to query the
`CONSTANT` state for the property.

Due to recent changes to the way `qmltypes` file are written, this
information is now exposed as `isPropertyConstant` in that and adjacent
contexts.

The private flag that is used to store the information for
set/`isConstant` was renamed to `m_isPropertyConstant` in a previous
patch, but the public interface was left unchanged.

To improve consistency, `setIsConstant` and `isConstant` are now renamed
to `setIsPropertyConstant` and `isPropertyConstant`.

The few usages of the methods were adapted to the change.

Change-Id: Ice431fe96ad360c98a1b786bebe87bbecac28125
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>