SelectionRectangle: clear selection from TableView
If the TableView delegate has a TapHandler, it will shadow
the TapHandler in TableView that normally clears the old
selection when a new drag selection begins. This means
that the SelectionRectangle will keep on extending the
current selection, even when no modifiers are being held.

This patch will make the selection logic more robust
by leaving it to QQuickTableView::startSelection() to
clear (or not clear) the existing selection. Then we
don't depend on the TapHandler in TableView to be called,
and we can also simplify SelectionRectangle a bit.

Pick-to: 6.8 6.7 6.6
Task-number: QTBUG-125867
Change-Id: Iaa3c6414d916773f62de4faec5a2fcb7b2f3e6d0
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>