Material: fix button padding for IconOnly and TextOnly displays
Correct the right padding of the Material style Button when the display
is IconOnly and the left padding when the display is TextOnly.

The issue was due to the buttonLeftPadding() and buttonRightPadding()
functions not considering the display condition in the hasIcon and
hasText arguments. To fix this, the display condition is now included in
the arguments of these functions.

Additionally, comparing the leftPadding and rightPadding with the output
of buttonLeftPadding() and buttonRightPadding() has been added to the
test case.

Fixes: QTBUG-126124
Change-Id: I2f1b60ebeae3d9854f96357bc5b0b5a77746aa17
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit bffee8b06d960d3c84fab46895ae2edbe1e066c1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 41a517ebf2704665f18405b466695b754411f468)
(cherry picked from commit 218653b05116b70b075a30520892e7c1c2d471fa)