Stabilize tst_QQuickPopup::popupWindowChangingParent
I had accidentally added a `visible: true` binding on the Popup,
in the QML document used by the test. Which caused the main window
to not always be initialized, before the popup window.

The main window's geometry is used by the QQuickItem::mapToGlobal()
function, which is called when comparing the global position of the
popup window's parent item, to verify that the popup window is
positioned relative to it.

Removing the `visible: true` binding, seems to fix the issue.
But further improvements to the test are also added to make it more
robust.

The following improvements are being made to the macros VERIFY_GLOBAL_POS
and VERIFY_LOCAL_POS, for the sake of making future debugging easier:
  - A better failure message.
  - They no longer need to store any local variables on the stack.
  - Redundant compare statements are removed.

Since the test is fixed, it's now safe to unblacklist.

Fixes: QTBUG-126175
Change-Id: I1032494818bf85bf8aaf3d2d60b56c0d991855b4
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit ef9238c99c25fc7e2d65b58eca2f2cb7ccf1c44d)
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>