Prepare for Q{AbstractAnimation,Unified}Timer::runningAnimationCount() return value change
Applying headersclean_check to private headers has pointed out that
QAbstractAnimation{Timer,Job}::runningAnimationCount() narrow
qsizetype to int in their respective return statements.

Since this is private API, we can change the return type of the
functions called runningAnimationCount() (incl. the one in
QUnifiedTimer) from int to qsizetype, but have to adapt the users in
other modules first.

QtDeclarative is the only such module.

For the virtual function, use a C++ macro defined in the respective
QtBase change.

For the caller, just receive as qsizetype already, and safely narrow
when passing to the next function, which still takes int.

Task-number: QTBUG-103530
Task-number: QTBUG-126219
Change-Id: Ib3d57290e74e8c099bbf23c615c409525ce690d2
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>