QQuickRt: Support picking a D24S8 alternative also on the multiview path
The ideal alternative is 32-bit depth + 8 bit stencil (+24 bit unused)
(a 64-bit format), but that has no equivalent in Qt 6.8.
From 6.9 on we have D32FS8 in QRhiTexture::Format. Use D32F in 6.8,
D32FS8 in 6.9. Won't make a difference given Quick 3D is not relying
on stencil.

Needs a dedicated patch for 6.8.

Change-Id: I6c7625cce459b74ef6742b8c72b97ddaa9a17260
Reviewed-by: Andy Nichols <andy.nichols@qt.io>