QQuickPopupWindow: set initial size
If you do:

Popup {
   width: 200
   height: 200
}

then QQuickPopup::setWidth/Height will be called
during construction, before a QQuickWindow is yet
created. Since we cannot resize the window at that,
point, we fall back to resize the popupItem instead.
But then it's important to remember to forward the
requested size to the window once we create it.

Pick-to: 6.8
Change-Id: I1e7dafe924f90dfb4e2580cc4b9418ecaf9653d6
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>