Make sure ToolTip's children palette is updated
Popups are not QQuickItems so when updating the palette, we explicitly
called QQuickPopupPrivate::updateContentPalettes for all children of
type QQuickPopup of the application window.
However, this didn't always work for tooltips when they were created
via attached objects (QQuickToolTipAttached), which is a common way to
create them. When QQuickToolTipAttached created the tooltip instance,
the parent of it wouldn't be the parent item, but the QML engine. In
this case we wouldn't find the toolTip in the list of children of the
application window.
To fix, we explicitly search for children of type QQuickToolTipAttached
instead, whose parent is the parent item.

Pick-to: 6.8 6.7 6.5
Change-Id: I70425b731f998622c93421eed11067a36b39fd75
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>