VectorImage qml generator: Add line continuation option
Allow disabling the automatic newline functionality of the internal
stream() function, so that the code can add more to the same line.

This changes the functionality so that a newline is added immediately
by stream(), instead of when the returned temp object is destructed.
I.e. now calling stream() starts a new line, except if the SameLine
flag is passed.

This also allows simplification of the whole streaming code, with
fewer intermediate data copies and allocations.

Pick-to: 6.8
Change-Id: I567efd96bd2cb86e73212208d5e7f9328251afa3
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>