JNI: prepare for standard Java types getting declared in qtjnitypes.h
Only declare the standard types locally if the
QT_DECLARE_JNI_CLASS_STANDARD_TYPES preprocessor symbol is undefined.
That allows us to merge the relevant changes in each submodule
independently, without having to orchestrate the submodule update.

Once all changes are in and the submodule update is through, remove the
declarations entirely.

Pick-to: 6.8
Change-Id: I88045c7d88a5f6f5ac00fe41389d63206ad854ad
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>