QuickDialogs: Explicitly link QuickTemplates2Private
This allows qmltyperegistrar to see the types from Templates and avoid
duplicating them in Dialogs. The linkage was there anyway, but it was a
private linkage from QuickControls2Impl before.

This removes the last duplication of QPlatformDialogHelper as found in
the QML API review.

Change-Id: Ia29355f934f0c54fe91b340104b3652a7e38a4e5
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 90bca64739609bea075629d3865243cdff4cd41d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>