QQuickPopup: remove PopupType::Default
PopupType::Default was meant as a way for a developer
to let Qt choose the best popup type, based on the
run-time (or build-time) environment.

But this could also be achieved by just setting the default value
of popupType already in the popup constructor (or when resetting
popupType) to either Item, Window, or Native, based on the same
information.

The difference is that the popupType cannot always be decided
before the popup is about to open. A menu, for example, can be
reparented at run-time into another menu, or into a native menubar.
And then the effective popup type needs to be the same as the parent.
But this will always be the case, even if the popupType is
explicitly set to something else than PopupType::Default.

So, at the end of the day, setting popupType to Window or Native
should be understood as a hint (or preference). And this might or might
not be respected, based on platform capabilities and the run-time state
of the application. And leaving it unspecified will give you
what is recommended on the platform by default, run-time considerations
aside. As a result, PopupType::Default can therefore be removed, since
it doesn't offer any extra logic on top of that.

Change-Id: I174ba70b6553e7fb867d68aa4593eb6cc23aefd3
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit f8b97e5af566f8c5badf1059d1b54055a250490d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>