dom: add representation for super and this literals
Add a dom representation for super and this literals for qmlls to not
bail out on files containing them.
Arguably, there is no need for ECMAscript classes in QML, so only add
minimal support for them (dom representation + disable completions on
them).

Also add tests to make sure the completions on them are disabled.

Task-number: QTBUG-92876
Change-Id: I9012bcd6a79a4610ff02e4ec474ca881faa5b7d6
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>