QmlCompiler: Require QQmlJSRegisterContent as scope types
This is the first step in chaining up the values in a graph. We can only
derive the next node of the graph if we have a full previous node.
QQmlJSRegisterContent will ultimately just be a container for a node.
Prepare for this by requiring QQmlJSRegisterContent as input when
creating a new register content.

In some cases we can't actually nest contents, yet. Therefore we chicken
out by creating synthetic QQmlJSRegisterContents.

Task-number: QTBUG-124670
Change-Id: Idad793692ca78d71ba17107d396e508bc1392b67
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>