AnchorChanges: Account for undefined opposite edge on state reset
When a state with AnchorChanges is reset, there is logic to check if the
state altered the target size in situations where an anchor is not
originally set, but is in the state. This check didn't take into account
the possibility that the opposite edge of the anchor set in state may
have been set to undefined, and the state didn't actually set the Item
size.

This patch adds an additional check for undefined opposite edge in state
to get rid of this false positive.

Fixes: QTBUG-126057
Pick-to: 6.7 6.5
Change-Id: I998c5da7db7d6fc3e50909ab6a1b4d972425d807
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 0232ef390ff44dfc2c057da74974eff3a344b92b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>