QQuickWidgets: Don't recreate root item on window change
Previously, QQuickWidgetPrivate::execute() was used in
QQuickWidgetPrivate::handleWindowChange to reset the state of the root
component and reparent it to the new window. This is only performed
if there is a source, presumably to create the new root element from the
latest available one. This however doesn't seem necessary as setting a
new source already calls execute() meaning the current root element is
already based on it. Therefore, it should be sufficient to simply
reparent.

This change fixes QTBUG-126136. When calling setLayout in MyWidget, the
root element is reparented and this leads to
qSendWindowChangeToTextureChildrenRecursively() and ultimatly
handleWindowChange() being called. There the root element get's deleted
and recreated within execute. Back in the user code, this has the
visible effect of breaking the connection that had been created as well
as leaving the pointer to the deleted root element dangling.

This problem was partially hidden in 6.5 and from 6.7.0 onwards by
1bd755465efa27294362925f55306f88f1800936 when it prevented
qSendWindowChangeToTextureChildrenRecursively() from being called from
QWidget::setParent().

Pick-to: 6.5
Fixes: QTBUG-126136
Change-Id: Id7c820a88f2c0265ea9a3c1e3f02a045356087d5
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 81ef787a6f41f99f889fd24be94ce0682ea7df75)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 37574f21f9d878c6711d4a9525488b68faf76267)