fix crash when multiline token includes \n character
qmlls was crashing when a multiline string literal contains raw newline
characters \n or \r because the method actually expects the code
in the text editor before processed by the lexer.

Use the literal value before interpreted by lexer to detect
the multiline tokens location in the editor. This is needed by
the semantic highlighting tool.

Pick-to: 6.8
Change-Id: I4b11aa0692aea13ae66d90639cf2996b4656ad3a
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>