qmlls: implement completion for new and new member expressions
Add completions and test for new and new member expressions. Arguably,
there is no need for ECMA script classes in QML, so only implement the
completion feature to not suggest weird stuff inside of new and new
member expressions and ignore the go-to-definition,
go-to-type-definition, find-usages and rename-usages feature for these
expressions.

Task-number: QTBUG-92876
Change-Id: Ib453d26a45969e33d7a51b7e1378126fa41f62ac
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>