dom: add test for metaproperty (new.target)
There is a special metaproperty `new.target` in JS that is currently
parsed like a normal field member access like `object.someProperty` for
example. Add a test in tst_qmldomitem.h to check that it gets the
correct dom construction as a ScriptBinaryExpression, and that it does
not need a special dom representation.

This is the last thing missing from QTBUG-92876.

Fixes: QTBUG-92876
Change-Id: I2e345a7f63a47fe1c61372bf64cf02cdd0cf6d9a
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 69fff1e9033891585bbe26cfd13bf7d9c39c7b0c)
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>