Unblacklist tst_qquickfontdialogimpl::settingUnderlineAndStrikeoutEffects
A font change in the font dialog will trigger a font change in the
sample text's textEdit. This may change the size of the text edit,
and require a relayout of the dialog's items.
That is what was happenning in tst_qquickfontdialogimpl::
settingUnderlineAndStrikeoutEffects and causing a failure to click a
button because the items hadn't finished the relayout. To fix, before
triggering any mouse events after the font change, we need to wait for
a polish event using qWaitForPolish.

Fixes: QTBUG-126375
Change-Id: Icb69a64bd27a78852b509213f792f2ee2f3bcf4f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>