Fix swipe to remove issue in swipe delegate
The global position used for calculating the drag distance causes an
issue when swipe delegates are placed at different axes. For instance,
a horizontal swipe can remove items even when swipe delegates are placed
vertically. This is because the calculated drag distance will be w.r.t
to the screen coordinates system and not the item coordinate system.

This patch fixes that issue by converting the screen coordinates to item
coordinates when calculating the drag distance.

Fixes: QTBUG-125906
Pick-to: 6.8 6.6 6.7 6.5
Change-Id: If6dd03d43e87386f807ec8bf8c16eb0566b79363
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>